# generated/node/src/routes/dashboards.routes.ts
// @generated by flusk-lang — DO NOT EDIT

import { type FastifyInstance } from 'fastify';
import type Database from 'better-sqlite3';
import { listDashboards } from '../functions/list-dashboards.function.js';
import { createDashboard } from '../functions/create-dashboard.function.js';
import { findDashboardById } from '../functions/find-dashboard-by-id.function.js';
import { updateDashboard } from '../functions/update-dashboard.function.js';
import { deleteDashboard } from '../functions/delete-dashboard.function.js';
import { renderDashboardData } from '../functions/render-dashboard-data.function.js';

interface AppWithDb extends FastifyInstance { db: Database.Database; }

export const dashboardsRoutes = async (app: FastifyInstance): Promise<void> => {
  const { db } = app as AppWithDb;

  app.get('/api/dashboards', async (request, reply) => {
    try {
      const result = listDashboards(db);
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'listDashboards failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.post('/api/dashboards', async (request, reply) => {
    try {
      const body = request.body as Record<string, unknown>;
      const result = createDashboard(db, body);
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'createDashboard failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.get('/api/dashboards/:id', async (request, reply) => {
    try {
      const { id } = request.params as Record<string, string>;
      const result = findDashboardById(db, id);
      if (result === null || result === undefined) {
        return reply.status(404).send({ error: 'Not found' });
      }
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'findDashboardById failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.put('/api/dashboards/:id', async (request, reply) => {
    try {
      const body = request.body as Record<string, unknown>;
      const result = updateDashboard(db, body);
      if (result === null || result === undefined) {
        return reply.status(404).send({ error: 'Not found' });
      }
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'updateDashboard failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.delete('/api/dashboards/:id', async (request, reply) => {
    try {
      const { id } = request.params as Record<string, string>;
      const result = deleteDashboard(db, id);
      return reply.status(204).send();
    } catch (err: unknown) {
      request.log.error(err, 'deleteDashboard failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.get('/api/dashboards/:id/render', async (request, reply) => {
    try {
      const { id } = request.params as Record<string, string>;
      const result = renderDashboardData(db, id);
      if (result === null || result === undefined) {
        return reply.status(404).send({ error: 'Not found' });
      }
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'renderDashboardData failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

};