# generated/node/src/routes/insights.routes.ts
// @generated by flusk-lang — DO NOT EDIT

import { type FastifyInstance } from 'fastify';
import type Database from 'better-sqlite3';
import { listInsights } from '../functions/list-insights.function.js';
import { findInsightById } from '../functions/find-insight-by-id.function.js';
import { updateInsight } from '../functions/update-insight.function.js';

interface AppWithDb extends FastifyInstance { db: Database.Database; }

export const insightsRoutes = async (app: FastifyInstance): Promise<void> => {
  const { db } = app as AppWithDb;

  app.get('/api/insights', async (request, reply) => {
    try {
      const result = listInsights(db);
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'listInsights failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.get('/api/insights/:id', async (request, reply) => {
    try {
      const { id } = request.params as Record<string, string>;
      const result = findInsightById(db, id);
      if (result === null || result === undefined) {
        return reply.status(404).send({ error: 'Not found' });
      }
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'findInsightById failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.put('/api/insights/:id/status', async (request, reply) => {
    try {
      const body = request.body as Record<string, unknown>;
      const result = updateInsight(db, body);
      if (result === null || result === undefined) {
        return reply.status(404).send({ error: 'Not found' });
      }
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'updateInsight failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

};