# generated/node/src/routes/llmCalls.routes.ts
// @generated by flusk-lang — DO NOT EDIT

import { type FastifyInstance } from 'fastify';
import type Database from 'better-sqlite3';
import { listLlmCalls } from '../functions/list-llm-calls.function.js';
import { createLlmCall } from '../functions/create-llm-call.function.js';
import { findLlmCallById } from '../functions/find-llm-call-by-id.function.js';
import { queryLlmCall } from '../functions/query-llm-call.function.js';
import { detectDuplicates } from '../functions/detect-duplicates.function.js';

interface AppWithDb extends FastifyInstance { db: Database.Database; }

export const llmCallsRoutes = async (app: FastifyInstance): Promise<void> => {
  const { db } = app as AppWithDb;

  app.get('/api/llm-calls', async (request, reply) => {
    try {
      const result = listLlmCalls(db);
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'listLlmCalls failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.post('/api/llm-calls', async (request, reply) => {
    try {
      const body = request.body as Record<string, unknown>;
      const result = createLlmCall(db, body);
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'createLlmCall failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.get('/api/llm-calls/:id', async (request, reply) => {
    try {
      const { id } = request.params as Record<string, string>;
      const result = findLlmCallById(db, id);
      if (result === null || result === undefined) {
        return reply.status(404).send({ error: 'Not found' });
      }
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'findLlmCallById failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.get('/api/llm-calls/aggregate/by-model', async (request, reply) => {
    try {
      const result = queryLlmCall(db);
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'queryLlmCall failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.get('/api/llm-calls/aggregate/by-agent', async (request, reply) => {
    try {
      const result = queryLlmCall(db);
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'queryLlmCall failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.get('/api/llm-calls/duplicates', async (request, reply) => {
    try {
      const result = detectDuplicates(db);
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'detectDuplicates failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

};