# generated/node/src/routes/optimizations.routes.ts
// @generated by flusk-lang — DO NOT EDIT

import { type FastifyInstance } from 'fastify';
import type Database from 'better-sqlite3';
import { listOptimizations } from '../functions/list-optimizations.function.js';
import { findOptimizationById } from '../functions/find-optimization-by-id.function.js';
import { updateOptimization } from '../functions/update-optimization.function.js';

interface AppWithDb extends FastifyInstance { db: Database.Database; }

export const optimizationsRoutes = async (app: FastifyInstance): Promise<void> => {
  const { db } = app as AppWithDb;

  app.get('/api/optimizations', async (request, reply) => {
    try {
      const result = listOptimizations(db);
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'listOptimizations failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.get('/api/optimizations/:id', async (request, reply) => {
    try {
      const { id } = request.params as Record<string, string>;
      const result = findOptimizationById(db, id);
      if (result === null || result === undefined) {
        return reply.status(404).send({ error: 'Not found' });
      }
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'findOptimizationById failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.put('/api/optimizations/:id/status', async (request, reply) => {
    try {
      const body = request.body as Record<string, unknown>;
      const result = updateOptimization(db, body);
      if (result === null || result === undefined) {
        return reply.status(404).send({ error: 'Not found' });
      }
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'updateOptimization failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

};