# generated/node/src/routes/partner-integrations.routes.ts
// @generated by flusk-lang — DO NOT EDIT

import { type FastifyInstance } from 'fastify';
import type Database from 'better-sqlite3';
import { listPartnerIntegration } from '../functions/list-partner-integration.function.js';
import { createPartnerIntegration } from '../functions/create-partner-integration.function.js';
import { findPartnerIntegrationById } from '../functions/find-partner-integration-by-id.function.js';
import { updatePartnerIntegration } from '../functions/update-partner-integration.function.js';
import { deletePartnerIntegration } from '../functions/delete-partner-integration.function.js';

interface AppWithDb extends FastifyInstance { db: Database.Database; }

export const partnerIntegrationsRoutes = async (app: FastifyInstance): Promise<void> => {
  const { db } = app as AppWithDb;

  app.get('/api/partner-integrations', async (request, reply) => {
    try {
      const result = listPartnerIntegration(db);
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'listPartnerIntegration failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.post('/api/partner-integrations', async (request, reply) => {
    try {
      const body = request.body as Record<string, unknown>;
      const result = createPartnerIntegration(db, body);
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'createPartnerIntegration failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.get('/api/partner-integrations/:id', async (request, reply) => {
    try {
      const { id } = request.params as Record<string, string>;
      const result = findPartnerIntegrationById(db, id);
      if (result === null || result === undefined) {
        return reply.status(404).send({ error: 'Not found' });
      }
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'findPartnerIntegrationById failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.put('/api/partner-integrations/:id', async (request, reply) => {
    try {
      const body = request.body as Record<string, unknown>;
      const result = updatePartnerIntegration(db, body);
      if (result === null || result === undefined) {
        return reply.status(404).send({ error: 'Not found' });
      }
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'updatePartnerIntegration failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.delete('/api/partner-integrations/:id', async (request, reply) => {
    try {
      const { id } = request.params as Record<string, string>;
      const result = deletePartnerIntegration(db, id);
      return reply.status(204).send();
    } catch (err: unknown) {
      request.log.error(err, 'deletePartnerIntegration failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

};