# generated/openapi.yaml
openapi: 3.1.0
info:
  title: Flusk API
  version: 0.1.0
paths:
...
  /api/dashboards/:
    get:
      operationId: listDashboards
      tags:
        - dashboards
      summary: List all dashboards
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dashboard'
    post:
      operationId: createDashboard
      tags:
        - dashboards
      summary: Create a new dashboard
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dashboard'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Dashboard'
  /api/dashboards/:id:
    get:
      operationId: findDashboardById
      tags:
        - dashboards
      summary: Get dashboard by ID
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dashboard'
    put:
      operationId: updateDashboard
      tags:
        - dashboards
      summary: Update a dashboard
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dashboard'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Dashboard'
    delete:
      operationId: deleteDashboard
      tags:
        - dashboards
      summary: Delete a dashboard
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dashboard'
...
  /api/insights/:id/status:
    put:
      operationId: updateInsight
      tags:
        - insights
      summary: Update insight status (accepted/dismissed)
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Insight'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Insight'
  /api/llm-calls/:
    get:
      operationId: listLlmCalls
      tags:
        - llmCalls
      summary: List LLM calls with filtering
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LlmCall'
    post:
      operationId: createLlmCall
      tags:
        - llmCalls
      summary: Record a new LLM call
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LlmCall'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LlmCall'
...
  /api/optimizations/:id/status:
    put:
      operationId: updateOptimization
      tags:
        - optimizations
      summary: Update optimization status
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Optimization'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Optimization'
...
  /api/partner-integrations/:
    get:
      operationId: listPartnerIntegration
      tags:
        - partner-integrations
      summary: List integrations
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartnerIntegration'
    post:
      operationId: createPartnerIntegration
      tags:
        - partner-integrations
      summary: Create a new integration
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartnerIntegration'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PartnerIntegration'
  /api/partner-integrations/:id:
    get:
      operationId: findPartnerIntegrationById
      tags:
        - partner-integrations
      summary: Get integration by ID
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartnerIntegration'
    put:
      operationId: updatePartnerIntegration
      tags:
        - partner-integrations
      summary: Update integration
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartnerIntegration'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PartnerIntegration'
    delete:
      operationId: deletePartnerIntegration
      tags:
        - partner-integrations
      summary: Delete integration
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartnerIntegration'
...
components:
  schemas:
...
    Dashboard:
      type: object
      properties:
        name:
          type: string
          description: Dashboard name
        description:
          type: string
          description: What this dashboard shows
        layout:
          type: object
          description: Grid layout config (widgets, positions, sizes)
        filters:
          type: object
          description: Default filters (time range, model, agent)
        refreshInterval:
          type: integer
          description: Auto-refresh interval in seconds
        isDefault:
          type: boolean
          description: Show on home page
        owner:
          type: string
          description: Who created this dashboard
      required:
        - name
        - description
        - layout
        - filters
        - refreshInterval
        - isDefault
        - owner
      description: Configurable dashboard for visualizing observability data
...
    Insight:
      type: object
      properties:
        insightType:
          $ref: '#/components/schemas/enum'
          description: Type of insight
          enum:
            - cost_reduction
            - performance
            - caching
            - model_switch
            - pattern_elimination
            - quality_improvement
        title:
          type: string
          description: Short insight title
        description:
          type: string
          description: Detailed explanation
        estimatedSavingsUsd:
          type: number
          description: Projected monthly savings
        estimatedSavingsPercent:
          type: number
          description: Projected savings as percentage
        priority:
          $ref: '#/components/schemas/enum'
          description: Priority level
          enum:
            - high
            - medium
            - low
        status:
          $ref: '#/components/schemas/enum'
          description: Insight status
          enum:
            - new
            - viewed
            - accepted
            - dismissed
        affectedModel:
          type: string
          description: Model this insight relates to
        affectedAgent:
          type: string
          description: Agent this insight relates to
        analyzeSessionId:
          type: string
          description: Analysis session that generated this insight
        evidence:
          type: object
          description: Supporting data
      required:
        - insightType
        - title
        - description
        - estimatedSavingsUsd
        - estimatedSavingsPercent
        - priority
        - status
        - affectedModel
        - affectedAgent
        - analyzeSessionId
        - evidence
      description: AI-generated optimization insight
    LlmCall:
      type: object
      properties:
        traceId:
          type: string
          description: Associated trace ID for distributed tracing
        spanId:
          type: string
          description: Associated span ID
        model:
          type: string
          description: Model name (e.g. gpt-4, claude-3-opus)
        provider:
          $ref: '#/components/schemas/enum'
          description: LLM provider
          enum:
            - openai
            - anthropic
            - google
            - mistral
            - cohere
            - local
            - other
        promptHash:
          type: string
          description: SHA-256 hash of the prompt for dedup detection
        promptTokens:
          type: integer
          description: Number of input tokens
        completionTokens:
          type: integer
          description: Number of output tokens
        totalTokens:
          type: integer
          description: Total tokens used
        costUsd:
          type: number
          description: Cost in USD
        durationMs:
          type: integer
          description: Response time in milliseconds
        agentName:
          type: string
          description: Name of the agent or service that made the call
        promptCategory:
          type: string
          description: Category for grouping (e.g. summarization, extraction)
        cached:
          type: boolean
          description: Whether this was a cache hit
        error:
          type: string
          description: Error message if the call failed
        metadata:
          type: object
          description: Additional context (headers, params, etc.)
      required:
        - traceId
        - spanId
        - model
        - provider
        - promptHash
        - promptTokens
        - completionTokens
        - totalTokens
        - costUsd
        - durationMs
        - agentName
        - promptCategory
        - cached
        - error
        - metadata
      description: Individual LLM API call with cost and token tracking
...
    Optimization:
      type: object
      properties:
        optimizationType:
          $ref: '#/components/schemas/enum'
          description: Type of code optimization
          enum:
            - caching
            - batching
            - model_switch
            - prompt_compression
            - deduplication
            - streaming
        title:
          type: string
          description: Short title
        description:
          type: string
          description: Detailed explanation of the optimization
        codeBefore:
          type: string
          description: Original code snippet
        codeAfter:
          type: string
          description: Suggested optimized code
        filePath:
          type: string
          description: File path in the user's project
        lineNumber:
          type: integer
          description: Line number in the file
        estimatedSavingsUsd:
          type: number
          description: Projected monthly savings
        status:
          $ref: '#/components/schemas/enum'
          description: Optimization status
          enum:
            - pending
            - accepted
            - rejected
            - applied
        analyzeSessionId:
          type: string
          description: Analysis session that generated this
      required:
        - optimizationType
        - title
        - description
        - codeBefore
        - codeAfter
        - filePath
        - lineNumber
        - estimatedSavingsUsd
        - status
        - analyzeSessionId
      description: Generated code optimization suggestion
...
    PartnerIntegration:
      type: object
      properties:
        partnerId:
          type: string
          description: Partner that owns this integration
        name:
          type: string
          description: Integration name
        type:
          $ref: '#/components/schemas/enum'
          description: Integration type
          enum:
            - model-provider
            - tool
            - data-source
            - notification
        config:
          type: object
          description: Integration configuration
        status:
          $ref: '#/components/schemas/enum'
          description: Integration status
          enum:
            - active
            - inactive
            - pending-review
        endpointUrl:
          type: string
          description: Integration endpoint URL
        documentation:
          type: string
          description: Documentation URL or markdown
      required:
        - partnerId
        - name
        - type
        - config
        - status
        - endpointUrl
        - documentation
      description: Integration provided by a partner (model, tool, data source, etc.)
...